Desktop/Programming/CodePaste/requirements.txt
# AutoTyper Application Requirements

# Backend
fastapi>=0.104.0
uvicorn>=0.24.0
pymongo>=4.6.0
pyjwt>=2.8.0
pydantic[email]>=2.5.0

# Desktop App
pyperclip>=1.8.2
keyboard>=0.13.5
requests>=2.31.0

# For building exe
pyinstaller>=6.0.0
